backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.2
aiosignal==1.3.1
anyio==4.3.0
asgiref==3.9.1
async-timeout==4.0.3
attrs==23.2.0
azure-core==1.30.1
azure-identity==1.16.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==43.0.3
Deprecated==1.2.14
Django==5.2.4
django-cas-ng==4.3.0
django-extensions==3.2.3
djangorestframework==3.15.2
drf-yasg==1.21.7
et-xmlfile==1.1.0
exceptiongroup==1.2.0
frozenlist==1.4.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.4
httpx==0.27.0
hyperframe==6.0.1
idna==3.7
importlib-metadata==6.11.0
inflection==0.5.1
ldap3==2.9.1
lxml==5.1.0
microsoft-kiota-abstractions==1.3.1
microsoft-kiota-authentication-azure==1.0.0
microsoft-kiota-http==1.3.1
msal==1.27.0
msal-extensions==1.1.0
msgraph-core==1.0.0
multidict==6.0.5
mysqlclient==2.2.4
numpy==2.1.2
openai==0.27.10
openpyxl==3.1.5
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
packaging==24.0
pandas==2.2.3
ping3==4.0.5
portalocker==2.8.2
pyasn1==0.5.1
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.1
pyodbc==4.0.39
python-cas==1.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.4
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
std-uritemplate==0.0.55
termcolor==2.4.0
tqdm==4.66.5
typing_extensions==4.10.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.2
wrapt==1.16.0
yarl==1.9.4
zipp==3.19.1
